Separated lti data collection and added Kaltura to the mix.
