+++ csvParser/index.js
@@ -1,4 +1,5 @@
-const snr = require('./appModes/snr.js')
+const snr = require('./appModes/snr')
+const lti = require('./appModes/lti')
 const path = require('path')
 const fs = require('fs')
 const got = require('got')
@@ -6,22 +7,6 @@ const parse = require('csv-parse/lib/sync')
 const inquirer = require('inquirer')
 const FormData = require('form-data')
 const form = new FormData()
-
-function generateSchoolData (externalTool, eTFRecords, aDRecords, schoolMap) {
-  const accountId = eTFRecords.find(
-    record => record.external_tool_activation_id === externalTool.id
-  ).course_account_id
-  const accountName = aDRecords
-    .find(record => record.id === accountId)
-    .name.split(' ')[0]
-  let schoolAndStudentCount = [1, 0]
-  if (schoolMap.has(accountName)) {
-    schoolAndStudentCount = schoolMap.get(accountName)
-    schoolAndStudentCount[0] = schoolAndStudentCount[0] + 1
-  }
-  schoolMap.set(accountName, schoolAndStudentCount)
-  externalTool.accountName = accountName
-}
 
 const appModes = {
   compareSis: async () => {
@@ -257,287 +242,7 @@ const appModes = {
       }
     }
   },
-  gatherMWKStatistics: async () => {
-    const {
-      relativeETDDataPath,
-      relativeEDataFolder,
-      relativeETFDataPath,
-      relativeADataPath
-    } = await inquirer.prompt([
-      {
-        type: 'input',
-        name: 'relativeETDDataPath',
-        message:
-          'Please enter the relative path name of the external tools dim file!',
-        default: 'dataFiles/1171-external_tool_activation_dim-00000-ccf628c9'
-      },
-      {
-        type: 'input',
-        name: 'relativeEDataFolder',
-        message:
-          'Please enter the relative folder name where the enrollments files are stored!',
-        default: 'dataFiles'
-      },
-      {
-        type: 'input',
-        name: 'relativeETFDataPath',
-        message:
-          'Please enter the relative path name of the external tools fact file!',
-        default: 'dataFiles/1171-external_tool_activation_fact-00000-78616e2a'
-      },
-      {
-        type: 'input',
-        name: 'relativeADataPath',
-        message: 'Please enter the relative path name of the account file!',
-        default: 'dataFiles/1171-account_dim-00000-4842795c'
-      }
-    ])
-
-    let input = fs.readFileSync(
-      path.resolve(__dirname, relativeETDDataPath),
-      'utf8'
-    )
-    const eTDRecords = parse(input, {
-      delimiter: '\t',
-      quote: false,
-      columns: [
-        'id',
-        'canvas_id',
-        'course_id',
-        'account_id',
-        'activation_target_type',
-        'url',
-        'name',
-        'description',
-        'workflow_state',
-        'privacy_level',
-        'created_at',
-        'updated_at',
-        'tool_id',
-        'selectable_all'
-      ],
-      skip_empty_lines: true
-    })
-
-    const filesInEnrollmentsDir = fs.readdirSync(
-      path.resolve(__dirname, relativeEDataFolder),
-      'utf8'
-    )
-    console.debug(`Files in enrollments directory: ${filesInEnrollmentsDir}`)
-    const enrollmentsFilenames = filesInEnrollmentsDir.filter(item =>
-      item.startsWith('1171-enrollment_dim')
-    )
-    console.debug(`Relevant enrollments files: ${enrollmentsFilenames}`)
-    const multipleERecords = []
-    for (const enrollmentsFilename of enrollmentsFilenames) {
-      input = fs.readFileSync(
-        path.resolve(
-          __dirname,
-          `${relativeEDataFolder}/${enrollmentsFilename}`
-        ),
-        'utf8'
-      )
-      multipleERecords.push(
-        parse(input, {
-          delimiter: '\t',
-          quote: false,
-          columns: [
-            'id',
-            'canvas_id',
-            'root_account_id',
-            'course_section_id',
-            'role_id',
-            'type',
-            'workflow_state',
-            'created_at',
-            'updated_at',
-            'start_at',
-            'end_at',
-            'completed_at',
-            'self_enrolled',
-            'sis_source_id',
-            'course_id',
-            'user_id',
-            'last_activity_at'
-          ],
-          skip_empty_lines: true
-        })
-      )
-    }
-
-    input = fs.readFileSync(
-      path.resolve(__dirname, relativeETFDataPath),
-      'utf8'
-    )
-    const eTFRecords = parse(input, {
-      delimiter: '\t',
-      quote: false,
-      columns: [
-        'external_tool_activation_id',
-        'course_id',
-        'account_id',
-        'root_account_id',
-        'enrollment_term_id',
-        'course_account_id'
-      ],
-      skip_empty_lines: true
-    })
-
-    input = fs.readFileSync(path.resolve(__dirname, relativeADataPath), 'utf8')
-    const aDRecords = parse(input, {
-      delimiter: '\t',
-      quote: false,
-      columns: [
-        'id',
-        'canvas_id',
-        'name',
-        'depth',
-        'workflow_state',
-        'parent_account',
-        'parent_account_id',
-        'grandparent_account',
-        'grandparent_account_id',
-        'root_account',
-        'root_account_id',
-        'subaccount1',
-        'subaccount1_id',
-        'subaccount2',
-        'subaccount2_id',
-        'subaccount3',
-        'subaccount3_id',
-        'subaccount4',
-        'subaccount4_id',
-        'subaccount5',
-        'subaccount5_id',
-        'subaccount6',
-        'subaccount6_id',
-        'subaccount7',
-        'subaccount7_id',
-        'subaccount8',
-        'subaccount8_id',
-        'subaccount9',
-        'subaccount9_id',
-        'subaccount10',
-        'subaccount10_id',
-        'subaccount11',
-        'subaccount11_id',
-        'subaccount12',
-        'subaccount12_id',
-        'subaccount13',
-        'subaccount13_id',
-        'subaccount14',
-        'subaccount14_id',
-        'subaccount15',
-        'subaccount15_id',
-        'sis_source_id'
-      ],
-      skip_empty_lines: true
-    })
-
-    // ToDo: Group data based on school (i.e. subaccount)
-    const moebiusRecords = []
-    const moebiusSchoolMap = new Map()
-
-    const wirisRecords = []
-    const wirisSchoolMap = new Map()
-
-    for (const externalTool of eTDRecords) {
-      if (
-        externalTool.activation_target_type === 'course' &&
-        externalTool.workflow_state === 'active'
-      ) {
-        if (externalTool.url.includes('kth-mobius.mapleserver.com')) {
-          generateSchoolData(
-            externalTool,
-            eTFRecords,
-            aDRecords,
-            moebiusSchoolMap
-          )
-          moebiusRecords.push(externalTool)
-        } else if (externalTool.url.includes('wiris.net')) {
-          generateSchoolData(
-            externalTool,
-            eTFRecords,
-            aDRecords,
-            wirisSchoolMap
-          )
-          wirisRecords.push(externalTool)
-        }
-      }
-    }
-
-    let moebiusNumberOfEnrolledUsers = 0
-    for (const mRecord of moebiusRecords) {
-      const uniqueStudentsInCourse = new Map()
-      for (const eRecords of multipleERecords) {
-        for (const eRecord of eRecords) {
-          if (
-            eRecord.course_id === mRecord.course_id &&
-            (eRecord.workflow_state === 'active' ||
-              eRecord.workflow_state === 'completed') &&
-            eRecord.type === 'StudentEnrollment'
-          ) {
-            uniqueStudentsInCourse.set(eRecord.user_id)
-          }
-        }
-      }
-      moebiusNumberOfEnrolledUsers += uniqueStudentsInCourse.size
-      const schoolAndStudentCount = moebiusSchoolMap.get(mRecord.accountName)
-      schoolAndStudentCount[1] =
-        schoolAndStudentCount[1] + uniqueStudentsInCourse.size
-    }
-
-    let wirisNumberOfEnrolledUsers = 0
-    for (const mRecord of wirisRecords) {
-      const uniqueStudentsInCourse = new Map()
-      for (const eRecords of multipleERecords) {
-        for (const eRecord of eRecords) {
-          if (
-            eRecord.course_id === mRecord.course_id &&
-            (eRecord.workflow_state === 'active' ||
-              eRecord.workflow_state === 'completed') &&
-            eRecord.type === 'StudentEnrollment'
-          ) {
-            uniqueStudentsInCourse.set(eRecord.user_id)
-          }
-        }
-      }
-      wirisNumberOfEnrolledUsers += uniqueStudentsInCourse.size
-      const schoolAndStudentCount = wirisSchoolMap.get(mRecord.accountName)
-      schoolAndStudentCount[1] =
-        schoolAndStudentCount[1] + uniqueStudentsInCourse.size
-    }
-
-    // Present data
-    // Möbius
-    console.info(`Number of courses with Möbius LTI: ${moebiusRecords.length}`)
-    console.info('Courses grouped by school:')
-    for (const [key, value] of moebiusSchoolMap) {
-      console.info(`${key} = ${value[0]}`)
-    }
-    console.info(
-      `Number of students exposed to Möbius: ${moebiusNumberOfEnrolledUsers}`
-    )
-    console.info('Students grouped by school:')
-    for (const [key, value] of moebiusSchoolMap) {
-      console.info(`${key} = ${value[1]}`)
-    }
-    // New segment
-    console.info('===')
-    // Wiris
-    console.info(`Number of courses with Wiris LTI: ${wirisRecords.length}`)
-    console.info('Courses grouped by school:')
-    for (const [key, value] of wirisSchoolMap) {
-      console.info(`${key} = ${value[0]}`)
-    }
-    console.info(
-      `Number of students exposed to Wiris: ${wirisNumberOfEnrolledUsers}`
-    )
-    console.info('Students grouped by school:')
-    for (const [key, value] of wirisSchoolMap) {
-      console.info(`${key} = ${value[1]}`)
-    }
-  },
+  gatherMWKStatistics: lti,
   gatherLEQStatistics: snr
 }
